La App sobrevivio a los cambios grandes

+++ app/src/main/java/ec/yasuodev/proyecto_movil/ui/core/products/AddProductViewModel.kt
@@ -76,7 +76,7 @@ class AddProductViewModel : ViewModel() {
                 id = generateUUID(),
                 name = name.value!!,
                 store = store,
-                price = price.value!!.toFloat(),
+                price = price.value!!.toDouble(),
                 stock = stock.value!!.toInt()
             )
             try {
